backend/pyproject.toml
[project]
name = "askdata"
version = "0.1.0"
description = "NL→SQL self-service analytics platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "sqlglot>=25.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "apscheduler>=3.10.4",
    "aiogram>=3.7.0",
    "aiosmtplib>=3.0.0",
    "matplotlib>=3.9.0",
    "numpy>=1.26.0",
    "sentence-transformers>=3.0.0",
    "scikit-learn>=1.5.0",
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.0",
    "gigachat>=0.1.32",
    "anthropic>=0.40.0",
    "pytz>=2024.1",
    "vosk>=0.3.45",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/askdata"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
